nop_bai_tap

+++ module4/ss10_session_cookie/exercise/exercise_1/src/main/java/com/container/repository/IProductRepository.java
@@ -3,5 +3,5 @@ package com.container.repository;
 import com.container.model.Product;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface IProductRepository extends JpaRepository<Product,Integer> {
+public interface IProductRepository extends JpaRepository<Product, Long> {
 }
