+++ module4/ss10_session_cookie/exercise/exercise_1/src/main/java/com/container/service/impl/ProductService.java
@@ -19,7 +19,7 @@ public class ProductService implements IProductService {
     }
 
     @Override
-    public Optional<Product> findById(int id) {
+    public Optional<Product> findById(Long id) {
         return productRepository.findById(id);
     }
 }
